LOOP-1200: alert persistence followups (#94)
* Add query using modification counter

* PR Feedback

* LOOP-1200: Dump Alert database into Issue Report

Added query by date, plus unit tests.

* PR Feedback, plus some rework/refactor

This takes @darin's suggestion, with some modifications :).  Now, the Filter is part of the Anchor, which makes it so anchoring continues to apply the filter as appropriate.

* LOOP-1200: Split alertIdentifier and managerIdentifier

After talking with Paul, we determined that we will need to be able to either filter or group alerts by device, so it would be better to keep managerIdentifier separate now so it can be used in data queries.