LOOP-940: Initial pass at new Settings Screen
So far, this is a pretty close approximation of the Settings screen as spec'd in LOOP-940.
It is not completely wired up; only the following are wired up:
- "Closed Loop" (dosingEnabled)
- alert permissions
- pump & cgm manager buttons

Note: This screen is currently hidden behind the "Alert Permissions" button in the "old" Settings UI.
Note also: because the old UI (SettingsTableViewController) still exists and has a lot of code logic in it, the way it
navigates to the device screens is by first dismissing and then presenting them.  This is temporary.