Only set ntpClient's logCallback if you actually can
Sometimes, I'm not sure how, you can get into a situation where the TrueTime framework is compiled without `logCallback` in place (it uses DEBUG_LOGGING as its compile flag, but it is not in our scope).
This is a safer route: checking to see if it actually can be set before trying to set it.