LOOP-3938: Fix multiple alerting if multiple conditions
Becky found this in testing: the fundamental reason was a @Published var was being modified multiple times and that issued more than one alert.  This fixes that by coalescing the update of that state all at once, and cleans up the alert issuing/retracting so it is a bit more obvious what's going on.